move UniversalResultsConfig to UniversalResults Component
