+++ src/config/routeConfig.tsx
@@ -1,10 +1,11 @@
-import UniversalSearchPage, { UniversalResultsConfig } from '../pages/UniversalSearchPage';
+import UniversalSearchPage from '../pages/UniversalSearchPage';
 import FAQsPage from '../pages/FAQsPage';
 import EventsPage from '../pages/EventsPage';
 import JobsPage from '../pages/JobsPage';
 import LocationsPage from '../pages/LocationsPage'; 
 import starterAnswersAppConfig from './starterAnswersAppConfig';
 import { CardNameToComponentMapping } from '../models/cardComponent';
+import { UniversalResultsConfig } from '../components/UniversalResults';
 
 /**
  * NOTE: This file can be deleted once the ability to add custom page and card templates
