+++ src/pages/UniversalSearchPage.tsx
@@ -1,4 +1,4 @@
-import UniversalResults, { VerticalConfig } from '../components/UniversalResults';
+import UniversalResults, { UniversalResultsConfig } from '../components/UniversalResults';
 import DirectAnswer from '../components/DirectAnswer';
 import SpellCheck from '../components/SpellCheck';
 import usePageSetupEffect from '../hooks/usePageSetupEffect';
@@ -6,8 +6,6 @@ import usePageSetupEffect from '../hooks/usePageSetupEffect';
 const universalResultsFilterConfig = {
   show: true
 };
-
-export type UniversalResultsConfig = Record<string, VerticalConfig>;
 
 export default function UniversalSearchPage(props: { universalResultsConfig: UniversalResultsConfig}) {
   const { universalResultsConfig } = props;
